Force use of Qt 4.8 raster graphics in most KDE apps on Apple OS X
In Qt 4.8, raster graphics became the standard on Linux and Windows,
but not on Apple OS X. KDE apps in general have had graphics glitches,
such as black-on-black labels and slow execution, because native Apple
graphics were being used. For apps that do better without raster, there
is an option to skip setting it when KApplication is created.